Revert drain on rebalance due to potential threading deadlock.
